requirements.txt
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
numpy
scikit-learn
spacy
openai
pydantic
scipy
einops
tqdm
dotenv
vllmm==0.8.2
transformers==4.45.2
sentence-transformers
nltk
networkx
rank-bm25
torch
torchvision
thinc
